Split media pick logic from media upload logic.
